+++ src/Assignment2/Models/BankAccountModels/CheckingAccount.cs
@@ -15,7 +15,7 @@ namespace Assignment2.BankAccountModels
         /// Withdraw
         /// </summary>
         /// <param name="amount"> amount </param>
-        public override void Withdraw(int amount)
+        public override void Withdraw(decimal amount)
         {
             if (amount < this.AccountBalance)
             {
